Allow the Emacs Lisp checker to run in a buffer without a file name
Closes GH-1507.  This isn't perfect because org mode temporary buffers don't
have a file name at all, so we don't know which buffer they came from.

The original code said this:

> Ensure that we only check buffers with a backing file.  For buffers
> without a backing file we cannot guarantee that file names in error
> messages are properly resolved, because `byte-compile-file' emits file
> names *relative to the directory of the checked file* instead of the
> working directory.  Hence our backwards-substitution will fail, because
> the checker process has a different base directory to resolve relative
> file names than the Flycheck code working on the buffer to check.

The problem is that we create a temp file in /tmp/, and the checker reports
errors relative to that, so we get "xyz" as a file name.  This doesn't match
what we expect ($PWD/xyz).  The fix here is to tell the ELisp compiler which
directory to report errors relative to.